try implementing higher/lower game

+++ api/src/main/java/com/tammeoja/higherlower/entities/Movie.java
@@ -14,7 +14,7 @@ public record Movie(
         String overview,
         BigDecimal popularity,
         LocalDate releaseDate,
-        int revenue,
+        long revenue,
         Double runtime,
         String tagline,
         String title,
